pyproject.toml
[project]
name = "sahay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.6",
    "npm>=0.1.1",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.0",
    "uvicorn>=0.40.0",
]
